refactor: extract character card combat stats
