+++ src/lib/characterPreview/ShowcaseDpsScore.tsx
@@ -1,8 +1,8 @@
 import { SettingOutlined, SwapOutlined, SyncOutlined } from '@ant-design/icons'
 import { Button, Card, ConfigProvider, Flex, Segmented } from 'antd'
 import type { GlobalToken } from 'antd/es/theme/interface'
+import { CharacterCardCombatStats } from 'lib/characterPreview/CharacterCardCombatStats'
 import { OverlayText, showcaseOutline } from 'lib/characterPreview/CharacterPreviewComponents'
-import { CharacterCardCombatStats } from 'lib/characterPreview/CharacterScoringSummary'
 import { useAsyncSimScoringExecution } from 'lib/characterPreview/CharacterStatSummary'
 import StatText from 'lib/characterPreview/StatText'
 import { CUSTOM_TEAM, DEFAULT_TEAM, SETTINGS_TEAM } from 'lib/constants/constants'
